chore: download config files from the right commit

+++ e2e/_suites/fleet/stand-alone.go
@@ -8,6 +8,7 @@ import (
 	"context"
 	"fmt"
 	"os"
+	"path"
 	"strings"
 	"time"
 
@@ -47,10 +48,18 @@ func (sats *StandAloneTestSuite) afterScenario() {
 	}
 
 	if _, err := os.Stat(sats.AgentConfigFilePath); err == nil {
-		os.Remove(sats.AgentConfigFilePath)
-		log.WithFields(log.Fields{
-			"path": sats.AgentConfigFilePath,
-		}).Debug("Elastic Agent configuration file removed.")
+		beatsLocalPath := shell.GetEnv("BEATS_LOCAL_PATH", "")
+		if beatsLocalPath == "" {
+			os.Remove(sats.AgentConfigFilePath)
+
+			log.WithFields(log.Fields{
+				"path": sats.AgentConfigFilePath,
+			}).Debug("Elastic Agent configuration file removed.")
+		} else {
+			log.WithFields(log.Fields{
+				"path": mts.configurationFile,
+			}).Debug("Elastic Agent configuration file not removed because it's part of a repository.")
+		}
 	}
 }
 
@@ -66,6 +75,9 @@ func (sats *StandAloneTestSuite) aStandaloneAgentIsDeployed(image string) error 
 
 	dockerImageTag := agentVersion
 
+	descriptorLocation := shell.Getenv("GITHUB_CHECK_SHA1", "master")
+	configurationFileURL := "https://raw.githubusercontent.com/elastic/beats/" + descriptorLocation + "/x-pack/elastic-agent/elastic-agent.docker.yml"
+
 	useCISnapshots := shell.GetEnvBool("BEATS_USE_CI_SNAPSHOTS")
 	beatsLocalPath := shell.GetEnv("BEATS_LOCAL_PATH", "")
 	if useCISnapshots || beatsLocalPath != "" {
@@ -79,6 +91,17 @@ func (sats *StandAloneTestSuite) aStandaloneAgentIsDeployed(image string) error 
 		dockerImageTag += "-amd64"
 	}
 
+	var configurationFilePath string
+	if beatsLocalPath != "" {
+		configurationFilePath = path.Join(beatsLocalPath, "x-pack", "elastic-agent", "elastic-agent.docker.yml")
+	} else {
+		p, downloadError := e2e.DownloadFile(configurationFileURL)
+		if downloadError != nil {
+			return downloadError
+		}
+		configurationFilePath = p
+	}
+
 	serviceManager := services.NewServiceManager()
 
 	profileEnv["elasticAgentDockerImageSuffix"] = ""
@@ -90,12 +113,6 @@ func (sats *StandAloneTestSuite) aStandaloneAgentIsDeployed(image string) error 
 
 	containerName := fmt.Sprintf("%s_%s_%d", FleetProfileName, ElasticAgentServiceName, 1)
 
-	configurationFileURL := "https://raw.githubusercontent.com/elastic/beats/master/x-pack/elastic-agent/elastic-agent.docker.yml"
-
-	configurationFilePath, err := e2e.DownloadFile(configurationFileURL)
-	if err != nil {
-		return err
-	}
 	sats.AgentConfigFilePath = configurationFilePath
 
 	profileEnv["elasticAgentContainerName"] = containerName
